+++ externs/uint-buffer.js
@@ -1,12 +1,35 @@
+/*
+ * Copyright (c) 2018-2019 Rafael da Silva Rocha.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ */
+
 /**
- * @fileoverview Externs for uint-buffer 0.1
+ * @fileoverview Externs for uint-buffer 1.0
  * @see https://github.com/rochars/uint-buffer
  * @externs
  */
 
 /**
- * A class to write and read two's complement ints and unsigned ints
- * to and from byte buffers.
+ * A class to write and read unsigned ints to and from byte buffers.
  */
 var UintBuffer = {};
 
@@ -54,7 +77,7 @@ UintBuffer.prototype.unpackUnsafe = function(buffer, index) {};
 /**
  * Throws error in case of overflow.
  * @param {number} num The number.
- * @throws {Error} on overflow.
+ * @throws {Error} On overflow.
  * @protected
  */
 UintBuffer.prototype.overflow = function(num) {};