+++ rollup.config.js
@@ -1,5 +1,25 @@
 /*
- * Copyright (c) 2018 Rafael da Silva Rocha.
+ * Copyright (c) 2018-2019 Rafael da Silva Rocha.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
  */
 
 /**
@@ -7,45 +27,26 @@
  * @see https://github.com/rochars/uint-buffer
  */
 
-import closure from 'rollup-plugin-closure-compiler-js';
-import fs from 'fs';
-
-// Externs
-const externsFile = fs.readFileSync('./externs/uint-buffer.js', 'utf8');
-
-// Legal
-const license = '/**\n'+
-' * @see https://github.com/rochars/uint-buffer\n' +
-' */\n';
-
-// GCC UMD footer, compatible with old browsers, Node and AMD loaders
-const footer = 
-  "var UintBuffer=exports;" +
-  "typeof module!=='undefined'?module.exports=exports :" +
-  "typeof define==='function'&&define.amd?define(['exports'],exports) :" +
-  "typeof global!=='undefined'?global.UintBuffer=exports:null;";
+import compiler from '@ampproject/rollup-plugin-closure-compiler';
 
 export default [
-  // UMD, minified
   {
-    input: 'uint-buffer.js',
+    input: 'index.js',
     output: [
       {
-        file: 'uint-buffer.umd.js',
-        format: 'cjs',
-        strict: false,
-        banner: 'var exports=exports||{};'
-      }
+        file: 'dist/uint-buffer.js',
+        name: 'UintBuffer',
+        format: 'umd'
+      },
     ],
     plugins: [
-      closure({
-        languageIn: 'ECMASCRIPT6',
-        languageOut: 'ECMASCRIPT5',
-        compilationLevel: 'ADVANCED',
-        warningLevel: 'VERBOSE',
-        outputWrapper: license + '%output%' + footer,
-        externs: [{src: externsFile + 'exports={};'}]
+      compiler({
+        language_in: 'ECMASCRIPT6',
+        language_out: 'ECMASCRIPT3',
+        compilation_level: 'SIMPLE',
+        warning_level: 'VERBOSE',
+        externs: ['externs/uint-buffer.js']
       })
     ]
-  },
+  }
 ];